split test class to feww smaller
